fix: solved general css
